pyproject.toml
[project]
name = "awesome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "dj-database-url>=2.3.0",
    "django>=5.1.6",
    "django-extensions>=3.2.3",
    "psycopg>=3.2.5",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest-django>=4.10.0",
    "ruff>=0.9.9",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"

testpaths = ["posts/test_views.py"]